LIBSEARCH-65. Coerce JSON "title" into string
The "title" value for an entry returned by WorldCat may not always
be a string. This was discovered accidentally when searching for
the term "music scores", and the first result has a title of "[1]",
which was return without the quotes in the JSON result from WorldCat.

When run through a JSON parser, the "[1]" was interpreted as an array
with a single element, and caused problems when treated like a string.

This fix simply coerces the "title" field into a string using the
"to_s" method, to ensure that the value will always be a string.

https://issues.umd.edu/browse/LIBSEARCH-65